+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - neuronignore: Exclude directories and markdown files via `.neuronignore`. (#499)
     - By default, recursively read Markdown files in notes directory. This behaviour can be disabled via the neuronignore plugin.
 - Reduce error verbosity (CLI & web) in the scenario of there being innumerous broken wiki-links.
+- `neuron rib` is renamed to `neuron gen` (neuron has migrated from rib/shake to reflex)
 
 ## Unreleased (v1 + v2)